Add BotUID field to Router. Update during events if empty
